ci: change output dir and add cargo make flow

+++ scripts/release/build.sh
@@ -4,20 +4,28 @@ set -xeo pipefail
 source .release-env
 
 # build the QIR SDK binary
-cargo build --bin ${SDK_BIN} --release --features llvm${LLVM_VERSION}-0
+if [[ $QIR_SDK_BUILD -eq 1 ]]; then
+    cargo build --bin ${SDK_BIN} --release --features llvm${LLVM_VERSION}-0
+else 
+    echo "Skipping QIR SDK build."
+fi
 
 # build the QCS C SDK shared library
-rm -rf ${TMP_DEPS_ABS_PATH}/qcs-sdk-c
-git clone https://github.com/rigetti/qcs-sdk-c ${TMP_DEPS_ABS_PATH}/qcs-sdk-c
-pushd ${TMP_DEPS_ABS_PATH}/qcs-sdk-c
-git fetch --all --tags
-git checkout tags/${C_SDK_VERSION} -b build-qir-sdk-${TAG}
-cargo build --release
-popd
+if [[ $C_SDK_BUILD -eq 1 ]]; then
+    rm -rf ${TMP_DEPS_ABS_PATH}/qcs-sdk-c
+    git clone https://github.com/rigetti/qcs-sdk-c ${TMP_DEPS_ABS_PATH}/qcs-sdk-c
+    pushd ${TMP_DEPS_ABS_PATH}/qcs-sdk-c
+    git fetch --all --tags
+    git checkout tags/${C_SDK_VERSION} -b build-qir-sdk-${TAG}
+    cargo build --release
+    popd
+else 
+    echo "Skipping C SDK build."
+fi
 
 # build the SDK helper library
 pushd helper
 cp helper.h helper.c
-clang -c -o libhelper.o helper.c -fPIE
+clang -c -o libhelper.o helper.c -fPIC
 clang -shared -L${TMP_DEPS_ABS_PATH}/qcs-sdk-c/target/release -lqcs -o libhelper.$LIB_EXT libhelper.o
 popd
